+++ types/src/components/GovBRCheckbox/GovBRCheckbox.stories.d.ts
@@ -1,9 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react';
-import { GovBRCheckbox } from './GovBRCheckbox';
-declare const meta: Meta<typeof GovBRCheckbox>;
-export default meta;
-type Story = StoryObj<typeof GovBRCheckbox>;
-export declare const Default: Story;
-export declare const States: Story;
-export declare const HiddenLabel: Story;
-//# sourceMappingURL=GovBRCheckbox.stories.d.ts.map