+++ types/src/components/GovBRFormCheckbox/GovBRFormCheckbox.stories.d.ts
@@ -1,5 +1,0 @@
-import type { Meta } from '@storybook/react';
-declare const _default: Meta;
-export default _default;
-export declare const WithFormValidation: () => import("react/jsx-runtime").JSX.Element;
-//# sourceMappingURL=GovBRFormCheckbox.stories.d.ts.map